fix(visset): Move Bake-able VisualizationSet to its own module
This should prevent possible errors when Grasshopper is not yet loaded.
